+++ lib/flutter_fritzapi.dart
@@ -2,4 +2,6 @@ library flutter_fritzapi;
 
 export 'package:flutter_fritzapi/flutter_fritzapi/model/devices.dart';
 export 'package:flutter_fritzapi/flutter_fritzapi/model/energy_stats.dart';
+export 'package:flutter_fritzapi/flutter_fritzapi/model/network_counters.dart';
+export 'package:flutter_fritzapi/flutter_fritzapi/model/wifi_client.dart';
 export 'package:flutter_fritzapi/flutter_fritzapi/fritz_api_client.dart';